requirements.txt
aiohttp==3.6.2
aniso8601==8.0.0
appnope==0.1.0
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
beautifulsoup4==4.9.1
bleach==3.1.4
blinker==1.4
cachelib==0.1
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.1
Flask-BabelEx==0.9.4
Flask-Cors==3.0.4
Flask-Dance==3.0.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.8
Flask-Security-Too==3.4.0
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
flaskfilemanager==0.0.6
geoip2==4.0.2
httplib2==0.17.0
idna==2.9
importlib-metadata==1.6.0
IPy==1.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
littlefish==0.0.56
lxml==4.5.2
MarkupSafe==1.1.1
maxminddb==2.0.2
mistune==0.8.4
multidict==4.7.6
nbconvert==5.6.1
nbformat==5.0.5
notebook==6.0.3
oauth2client==4.1.3
oauthlib==3.1.0
pandocfilters==1.4.2
parso==0.6.2
passlib==1.7.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.7.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
PyMarkovChain==1.8
PyMySQL==0.9.3
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
pyzmq==19.0.0
qtconsole==4.7.2
QtPy==1.9.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
Send2Trash==1.5.0
six==1.14.0
soupsieve==2.0.1
speaklater==1.3
SQLAlchemy==1.3.4
SQLAlchemy-Utils==0.36.3
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
typing-extensions==3.7.4.2
urllib3==1.25.8
URLObject==2.4.3
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
WTForms==2.2.1
yarl==1.5.1
zipp==3.1.0
